Move transaction related functions to their own action and api files
