+++ src/actions/account.test.js
@@ -1,14 +1,20 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import actionTypes from '../constants/actions';
-import { accountUpdated, accountLoggedOut,
-  secondPassphraseRegistered, delegateRegistered, sent, removePassphrase, passphraseUsed, loadDelegate } from './account';
+import {
+  accountUpdated,
+  accountLoggedOut,
+  secondPassphraseRegistered,
+  delegateRegistered,
+  removePassphrase,
+  passphraseUsed,
+  loadDelegate,
+} from './account';
 import { errorAlertDialogDisplayed } from './dialog';
 import { delegateRegisteredFailure } from './delegate';
 import * as accountApi from '../utils/api/account';
 import * as delegateApi from '../utils/api/delegate';
 import Fees from '../constants/fees';
-import { toRawLsk } from '../utils/lsk';
 import transactionTypes from '../constants/transactionTypes';
 import networks from '../constants/networks';
 import accounts from '../../test/constants/accounts';
@@ -187,79 +193,6 @@ describe('actions: account', () => {
         type: actionTypes.updateDelegate,
       };
       expect(dispatch).to.have.been.calledWith(updateDelegateAction);
-    });
-  });
-
-  describe('sent', () => {
-    let accountApiMock;
-    const data = {
-      activePeer: {},
-      recipientId: '15833198055097037957L',
-      amount: 100,
-      passphrase: 'sample passphrase',
-      secondPassphrase: null,
-      account: {
-        publicKey: 'test_public-key',
-        address: 'test_address',
-      },
-    };
-    const actionFunction = sent(data);
-    let dispatch;
-
-    beforeEach(() => {
-      accountApiMock = sinon.stub(accountApi, 'send');
-      dispatch = sinon.spy();
-    });
-
-    afterEach(() => {
-      accountApiMock.restore();
-    });
-
-    it('should create an action function', () => {
-      expect(typeof actionFunction).to.be.deep.equal('function');
-    });
-
-    it('should dispatch transactionAdded action if resolved', () => {
-      accountApiMock.returnsPromise().resolves({ transactionId: '15626650747375562521' });
-      const expectedAction = {
-        id: '15626650747375562521',
-        senderPublicKey: 'test_public-key',
-        senderId: 'test_address',
-        recipientId: data.recipientId,
-        amount: toRawLsk(data.amount),
-        fee: Fees.send,
-        type: transactionTypes.send,
-      };
-
-      actionFunction(dispatch);
-      expect(dispatch).to.have.been
-        .calledWith({ data: expectedAction, type: actionTypes.transactionAdded });
-    });
-
-    it('should dispatch transactionFailed action if caught', () => {
-      accountApiMock.returnsPromise().rejects({ message: 'sample message' });
-
-      actionFunction(dispatch);
-      const expectedAction = {
-        data: {
-          errorMessage: 'sample message.',
-        },
-        type: actionTypes.transactionFailed,
-      };
-      expect(dispatch).to.have.been.calledWith(expectedAction);
-    });
-
-    it('should dispatch transactionFailed action if caught but no message returned', () => {
-      accountApiMock.returnsPromise().rejects({});
-
-      actionFunction(dispatch);
-      const expectedAction = {
-        data: {
-          errorMessage: 'An error occurred while creating the transaction.',
-        },
-        type: actionTypes.transactionFailed,
-      };
-      expect(dispatch).to.have.been.calledWith(expectedAction);
     });
   });
