+++ src/actions/transactions.test.js
@@ -2,14 +2,53 @@ import { expect } from 'chai';
 import sinon from 'sinon';
 import actionTypes from '../constants/actions';
 import txFilters from './../constants/transactionFilters';
-import { transactionsRequested, loadTransaction } from './transactions';
-import * as accountApi from '../utils/api/account';
+import { sent, transactionsRequested, loadTransaction, transactionsUpdated } from './transactions';
+import * as transactionsApi from '../utils/api/transactions';
 import * as delegateApi from '../utils/api/delegate';
 import accounts from '../../test/constants/accounts';
+import transactionTypes from '../constants/transactionTypes';
+import Fees from '../constants/fees';
+import { toRawLsk } from '../utils/lsk';
 
 describe('actions: transactions', () => {
+  describe('transactionsUpdated', () => {
+    let transactionsApiMock;
+    const data = {
+      activePeer: {},
+      address: '15626650747375562521',
+      limit: 20,
+      offset: 0,
+      filter: txFilters.all,
+    };
+    const actionFunction = transactionsUpdated(data);
+    let dispatch;
+
+    beforeEach(() => {
+      transactionsApiMock = sinon.stub(transactionsApi, 'getTransactions');
+      dispatch = sinon.spy();
+    });
+
+    afterEach(() => {
+      transactionsApiMock.restore();
+    });
+
+    it('should dispatch transactionsLoaded action if resolved', () => {
+      transactionsApiMock.returnsPromise().resolves({ transactions: [], count: '0' });
+      const expectedAction = {
+        count: 0,
+        confirmed: [],
+      };
+
+      actionFunction(dispatch);
+      expect(dispatch).to.have.been.calledWith({
+        data: expectedAction,
+        type: actionTypes.transactionsUpdated,
+      });
+    });
+  });
+
   describe('transactionsRequested', () => {
-    let accountApiMock;
+    let transactionsApiMock;
     const data = {
       activePeer: {},
       address: '15626650747375562521',
@@ -21,12 +60,12 @@ describe('actions: transactions', () => {
     let dispatch;
 
     beforeEach(() => {
-      accountApiMock = sinon.stub(accountApi, 'transactions');
+      transactionsApiMock = sinon.stub(transactionsApi, 'getTransactions');
       dispatch = sinon.spy();
     });
 
     afterEach(() => {
-      accountApiMock.restore();
+      transactionsApiMock.restore();
     });
 
     it('should create an action function', () => {
@@ -34,7 +73,7 @@ describe('actions: transactions', () => {
     });
 
     it('should dispatch transactionsLoaded action if resolved', () => {
-      accountApiMock.returnsPromise().resolves({ transactions: [], count: '0' });
+      transactionsApiMock.returnsPromise().resolves({ transactions: [], count: '0' });
       const expectedAction = {
         count: 0,
         confirmed: [],
@@ -49,7 +88,7 @@ describe('actions: transactions', () => {
   });
 
   describe('loadTransaction', () => {
-    let accountApiMock;
+    let transactionApiMock;
     let delegateApiMock;
     const data = {
       activePeer: {
@@ -66,13 +105,13 @@ describe('actions: transactions', () => {
     let dispatch;
 
     beforeEach(() => {
-      accountApiMock = sinon.stub(accountApi, 'transaction');
+      transactionApiMock = sinon.stub(transactionsApi, 'getSingleTransaction');
       delegateApiMock = sinon.stub(delegateApi, 'getDelegate');
       dispatch = sinon.spy();
     });
 
     afterEach(() => {
-      accountApiMock.restore();
+      transactionApiMock.restore();
       delegateApiMock.restore();
     });
 
@@ -83,7 +122,7 @@ describe('actions: transactions', () => {
     it('should dispatch one transactionAddDelegateName action when transaction contains one vote added', () => {
       const delegateResponse = { delegate: { username: 'peterpan' } };
       const transactionResponse = { transaction: { votes: { added: [accounts.delegate.publicKey] }, count: '0' } };
-      accountApiMock.returnsPromise().resolves(transactionResponse);
+      transactionApiMock.returnsPromise().resolves(transactionResponse);
       delegateApiMock.returnsPromise().resolves(delegateResponse);
       const expectedActionPayload = {
         ...delegateResponse,
@@ -100,7 +139,7 @@ describe('actions: transactions', () => {
     it('should dispatch one transactionAddDelegateName action when transaction contains one vote deleted', () => {
       const delegateResponse = { delegate: { username: 'peterpan' } };
       const transactionResponse = { transaction: { votes: { deleted: [accounts.delegate.publicKey] }, count: '0' } };
-      accountApiMock.returnsPromise().resolves(transactionResponse);
+      transactionApiMock.returnsPromise().resolves(transactionResponse);
       delegateApiMock.returnsPromise().resolves(delegateResponse);
       const expectedActionPayload = {
         ...delegateResponse,
@@ -114,4 +153,82 @@ describe('actions: transactions', () => {
         .calledWith({ data: expectedActionPayload, type: actionTypes.transactionAddDelegateName });
     });
   });
+
+  describe('sent', () => {
+    let transactionsApiMock;
+    const data = {
+      activePeer: {},
+      recipientId: '15833198055097037957L',
+      amount: 100,
+      passphrase: 'sample passphrase',
+      secondPassphrase: null,
+      account: {
+        publicKey: 'test_public-key',
+        address: 'test_address',
+      },
+    };
+    const actionFunction = sent(data);
+    let dispatch;
+
+    beforeEach(() => {
+      transactionsApiMock = sinon.stub(transactionsApi, 'send');
+      dispatch = sinon.spy();
+    });
+
+    afterEach(() => {
+      transactionsApiMock.restore();
+    });
+
+    it('should create an action function', () => {
+      expect(typeof actionFunction).to.be.deep.equal('function');
+    });
+
+    it('should dispatch transactionAdded action if resolved', () => {
+      transactionsApiMock.returnsPromise().resolves({ transactionId: '15626650747375562521' });
+      const expectedAction = {
+        id: '15626650747375562521',
+        senderPublicKey: 'test_public-key',
+        senderId: 'test_address',
+        recipientId: data.recipientId,
+        amount: toRawLsk(data.amount),
+        fee: Fees.send,
+        type: transactionTypes.send,
+      };
+
+      actionFunction(dispatch);
+      expect(dispatch).to.have.been
+        .calledWith({ data: expectedAction, type: actionTypes.transactionAdded });
+    });
+
+    it('should dispatch transactionFailed action if caught', () => {
+      transactionsApiMock.returnsPromise().rejects({ message: 'sample message' });
+
+      actionFunction(dispatch);
+      const expectedAction = {
+        data: {
+          errorMessage: 'sample message.',
+        },
+        type: actionTypes.transactionFailed,
+      };
+      expect(dispatch).to.have.been.calledWith(expectedAction);
+    });
+
+    it('should dispatch transactionFailed action if caught but no message returned', () => {
+      transactionsApiMock.returnsPromise().rejects({});
+
+      actionFunction(dispatch);
+      const expectedAction = {
+        data: {
+          errorMessage: 'An error occurred while creating the transaction.',
+        },
+        type: actionTypes.transactionFailed,
+      };
+      expect(dispatch).to.have.been.calledWith(expectedAction);
+    });
+  });
+
+
+  // describe('accountLoggedOut', () => {
+  //   it('should create an action to reset the account', () => {
+  // });
 });
