+++ src/components/transactions/explorerTransactions/index.test.js
@@ -8,6 +8,7 @@ import { BrowserRouter as Router } from 'react-router-dom';
 import { prepareStore } from '../../../../test/utils/applicationInit';
 import * as accountAPI from '../../../../src/utils/api/account';
 import * as delegateAPI from '../../../../src/utils/api/delegate';
+import * as transactionsAPI from '../../../../src/utils/api/transactions';
 import peersReducer from '../../../store/reducers/peers';
 import accountReducer from '../../../store/reducers/account';
 import transactionReducer from '../../../store/reducers/transaction';
@@ -44,8 +45,8 @@ describe('ExplorerTransactions Component', () => {
   }, [thunk]);
 
   beforeEach(() => {
-    transactionsActionStub = stub(accountAPI, 'transactions');
-    transactionActionStub = stub(accountAPI, 'transaction');
+    transactionsActionStub = stub(transactionsAPI, 'getTransactions');
+    transactionActionStub = stub(transactionsAPI, 'getSingleTransaction');
     accountStub = stub(accountAPI, 'getAccount');
     delegateStub = stub(delegateAPI, 'getDelegate');
     delegateVotesStub = stub(delegateAPI, 'getVotes');
