+++ src/components/transactions/walletTransactions/index.test.js
@@ -6,8 +6,8 @@ import { expect } from 'chai';
 import { Provider } from 'react-redux';
 import { BrowserRouter as Router } from 'react-router-dom';
 import { prepareStore } from '../../../../test/utils/applicationInit';
-import * as accountAPI from '../../../../src/utils/api/account';
 import * as delegateAPI from '../../../../src/utils/api/delegate';
+import * as transactionsAPI from '../../../../src/utils/api/transactions';
 import peersReducer from '../../../store/reducers/peers';
 import accountReducer from '../../../store/reducers/account';
 import transactionsReducer from '../../../store/reducers/transactions';
@@ -40,7 +40,7 @@ describe('WalletTransactions Component', () => {
   }, [thunk]);
 
   beforeEach(() => {
-    transactionsActionsStub = stub(accountAPI, 'transactions');
+    transactionsActionsStub = stub(transactionsAPI, 'getTransactions');
     delegateVotesStub = stub(delegateAPI, 'getVotes');
     delegateVotersStub = stub(delegateAPI, 'getVoters');
