+++ src/utils/api/account.test.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 import { mock } from 'sinon';
-import { getAccount, setSecondPassphrase, send, transactions, unconfirmedTransactions } from './account';
+import { getAccount, setSecondPassphrase } from './account';
 
 describe('Utils: Account API', () => {
   const address = '1449310910991872227L';
@@ -50,27 +50,6 @@ describe('Utils: Account API', () => {
   describe('setSecondPassphrase', () => {
     it('should return a promise', () => {
       const promise = setSecondPassphrase();
-      expect(typeof promise.then).to.be.equal('function');
-    });
-  });
-
-  describe('send', () => {
-    it('should return a promise', () => {
-      const promise = send();
-      expect(typeof promise.then).to.be.equal('function');
-    });
-  });
-
-  describe('transactions', () => {
-    it('should return a promise', () => {
-      const promise = transactions({ activePeer: {} });
-      expect(typeof promise.then).to.be.equal('function');
-    });
-  });
-
-  describe('unconfirmedTransactions', () => {
-    it('should return a promise', () => {
-      const promise = unconfirmedTransactions();
       expect(typeof promise.then).to.be.equal('function');
     });
   });
